refactor: rename capturer/index.js
